refactor: RacingCarGame.java 패키지 변경
- controller에 해당하는 클래스는 racingcar 패키지의 루트로 이동하였습니다.
